#16 WiP

+++ sql/ws/ws_gw_trg_edit_node.sql
@@ -1,12 +1,12 @@
 /*
-This file is part of Giswater
+This file is part of Giswater 2.0
 The program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
 This version of Giswater is provided by Giswater Association
 */
 
 
    
-CREATE OR REPLACE FUNCTION SCHEMA_NAME.gw_trg_edit_node() RETURNS trigger LANGUAGE plpgsql AS $$
+CREATE OR REPLACE FUNCTION "SCHEMA_NAME".gw_trg_edit_node() RETURNS trigger LANGUAGE plpgsql AS $$
 DECLARE 
     inp_table varchar;
     man_table varchar;
@@ -29,30 +29,30 @@ BEGIN
         -- Node Catalog ID
         IF (NEW.nodecat_id IS NULL) THEN
             IF ((SELECT COUNT(*) FROM cat_node) = 0) THEN
-                RAISE EXCEPTION 'There are no nodes catalog defined in the model, define at least one.';
-            END IF;   
+                RAISE EXCEPTION '[%]: There are no nodes catalog defined in the model, define at least one.', TG_NAME;
+            END IF;      
         END IF;
 
         -- Sector ID
         IF (NEW.sector_id IS NULL) THEN
             IF ((SELECT COUNT(*) FROM sector) = 0) THEN
-                RAISE EXCEPTION 'There are no sectors defined in the model, define at least one.';
+                RAISE EXCEPTION '[%]: There are no sectors defined in the model, define at least one.', TG_NAME;
             END IF;
-            NEW.sector_id := (SELECT sector_id FROM sector WHERE (NEW.the_geom @ sector.the_geom) LIMIT 1);
+            NEW.sector_id:= (SELECT sector_id FROM sector WHERE (NEW.the_geom @ sector.the_geom) LIMIT 1);
             IF (NEW.sector_id IS NULL) THEN
-                RAISE EXCEPTION 'Please take a look on your map and use the approach of the sectors!!!';
-            END IF;
+                RAISE EXCEPTION '[%]: Please take a look on your map and use the approach of the sectors!', TG_NAME;
+            END IF;            
         END IF;
         
         -- Dma ID
         IF (NEW.dma_id IS NULL) THEN
             IF ((SELECT COUNT(*) FROM dma) = 0) THEN
-                RAISE EXCEPTION 'There are no dma defined in the model, define at least one.';
+                RAISE EXCEPTION '[%]: There are no dma defined in the model, define at least one.', TG_NAME;
             END IF;
             NEW.dma_id := (SELECT dma_id FROM dma WHERE (NEW.the_geom @ dma.the_geom) LIMIT 1);
             IF (NEW.dma_id IS NULL) THEN
-                RAISE EXCEPTION 'Please take a look on your map and use the approach of the dma!!!';
-            END IF;
+                RAISE EXCEPTION '[%]: Please take a look on your map and use the approach of the dma!', TG_NAME;
+            END IF;            
         END IF;
         
         -- FEATURE INSERT
@@ -73,9 +73,10 @@ BEGIN
 
         -- MANAGEMENT INSERT
         man_table:= (SELECT node_type.man_table FROM node_type JOIN cat_node ON (((node_type.id)::text = (cat_node.nodetype_id)::text)) WHERE cat_node.id=NEW.nodecat_id);
-        v_sql:= 'INSERT INTO '||man_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
-        EXECUTE v_sql;
-            
+        IF man_table IS NOT NULL THEN
+            v_sql:= 'INSERT INTO '||man_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
+            EXECUTE v_sql;
+        END IF;
         RETURN NEW;
 
 
@@ -96,8 +97,10 @@ BEGIN
             ELSIF (OLD.epa_type = 'PUMP') THEN
                 inp_table:= 'inp_pump';  
             END IF;
-            v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
-            EXECUTE v_sql;
+            IF inp_table IS NOT NULL THEN
+                v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
+                EXECUTE v_sql;
+            END IF;
 
             IF (NEW.epa_type = 'JUNCTION') THEN
                 inp_table:= 'inp_junction';   
@@ -112,16 +115,18 @@ BEGIN
             ELSIF (NEW.epa_type = 'PUMP') THEN
                 inp_table:= 'inp_pump';  
             END IF;
-            v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
+            IF inp_table IS NOT NULL THEN
+                v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
+            END IF;
             EXECUTE v_sql;
 
         END IF;
 
-        IF (NEW.nodecat_id <> OLD.nodecat_id) THEN  
-            old_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id)::text = (cat_node.nodetype_id)::text)) WHERE cat_node.id=OLD.nodecat_id)::text;
-            new_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id)::text = (cat_node.nodetype_id)::text)) WHERE cat_node.id=NEW.nodecat_id)::text;
-            IF (quote_literal(old_nodetype)::text <> quote_literal(new_nodetype)::text) THEN
-                RAISE EXCEPTION 'Change node catalog is forbidden. The new node catalog is not included on the same type (node_type.type) of the old node catalog';
+        IF (OLD.nodecat_id IS NOT NULL) AND (NEW.nodecat_id <> OLD.nodecat_id) THEN  
+            old_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id) = (cat_node.nodetype_id))) WHERE cat_node.id=OLD.nodecat_id);
+            new_nodetype:= (SELECT node_type.type FROM node_type JOIN cat_node ON (((node_type.id) = (cat_node.nodetype_id))) WHERE cat_node.id=NEW.nodecat_id);
+            IF (quote_literal(old_nodetype) <> quote_literal(new_nodetype)) THEN
+                RAISE EXCEPTION '[%]: Change node catalog is forbidden. The new node catalog is not included on the same type (node_type.type) of the old node catalog', TG_NAME;
             END IF;
         END IF;
 
@@ -144,11 +149,14 @@ BEGIN
     END IF;
 
 END;
-$$;
+$BODY$
+  LANGUAGE 'plpgsql' VOLATILE COST 100
+;
 
 
 
-CREATE TRIGGER gw_trg_edit_node INSTEAD OF INSERT OR DELETE OR UPDATE ON SCHEMA_NAME.v_edit_node
-FOR EACH ROW EXECUTE PROCEDURE SCHEMA_NAME.gw_trg_edit_node();
 
+
+CREATE TRIGGER gw_trg_edit_node INSTEAD OF INSERT OR DELETE OR UPDATE ON "SCHEMA_NAME".v_edit_node
+FOR EACH ROW EXECUTE PROCEDURE "SCHEMA_NAME".gw_trg_edit_node();
